Tests: move outdated tests to .oudated dir
