+++ tests/src/.outdated/addToContractAllowList.test.ts
@@ -16,13 +16,13 @@
 
 import chai from 'chai';
 import chaiAsPromised from 'chai-as-promised';
-import usingApi, {submitTransactionAsync, submitTransactionExpectFailAsync} from './substrate/substrate-api';
+import usingApi, {submitTransactionAsync, submitTransactionExpectFailAsync} from '../substrate/substrate-api';
 import {
   deployFlipper,
-} from './util/contracthelpers';
+} from '../util/contracthelpers';
 import {
   getGenericResult,
-} from './util/helpers';
+} from '../util/helpers';
 
 chai.use(chaiAsPromised);
 const expect = chai.expect;
