+++ tests/src/.outdated/collision-tests/adminDestroyCollection.test.ts
