+++ tests/src/.outdated/collision-tests/adminRightsOff.test.ts
