+++ tests/src/.outdated/collision-tests/sponsorPayments.test.ts
