+++ tests/src/.outdated/setChainLimits.test.ts
@@ -15,13 +15,13 @@
 // along with Unique Network. If not, see <http://www.gnu.org/licenses/>.
 
 import {IKeyringPair} from '@polkadot/types/types';
-import usingApi from './substrate/substrate-api';
+import usingApi from '../substrate/substrate-api';
 import {
   createCollectionExpectSuccess,
   addCollectionAdminExpectSuccess,
   setChainLimitsExpectFailure,
   IChainLimits,
-} from './util/helpers';
+} from '../util/helpers';
 
 // todo:playgrounds skipped ~ postponed
 describe.skip('Negative Integration Test setChainLimits', () => {
